requirements.txt
matplotlib
matplotlib-inline
kubernetes
python-dateutil
PyYAML
